requirements.txt
pyyaml
tornado
tornado-redis

